refactor: split Zarr functionality into multiple files
